+++ chrpc/Makefile
@@ -7,12 +7,13 @@ _SRCS		:= channel.c \
 			   channel_helpers.c \
 			   channel_local.c \
 			   channel_local2.c \
-			   serial.c
+			   serial_type.c \
+			   serial_value.c
 
 _TEST_SRCS   := main.c \
 				channel.c \
 				channel_local.c \
 				channel_local2.c \
-				serial.c
+				serial_type.c
 
 include ../lib_builder_stub.mk
