+++ chrpc/test/main.c
@@ -2,7 +2,7 @@
 #include "channel_local2.h"
 #include "chsys/sys.h"
 
-#include "serial.h"
+#include "serial_type.h"
 #include "unity/unity.h"
 #include "unity/unity_internals.h"
 
@@ -23,6 +23,6 @@ int main(void) {
     UNITY_BEGIN();
     channel_local_tests();
     channel_local2_tests();
-    chrpc_serial_tests();
+    chrpc_serial_type_tests();
     safe_exit(UNITY_END());
 }
