--- chrpc/test/serial.h
@@ -1,7 +1,0 @@
-
-#ifndef TEST_CHRPC_SERIAL_H
-#define TEST_CHRPC_SERIAL_H
-
-void chrpc_serial_tests(void);
-
-#endif
