+++ chrpc/test/serial_type.h
@@ -1,0 +1,7 @@
+
+#ifndef TEST_CHRPC_SERIAL_TYPE_H
+#define TEST_CHRPC_SERIAL_TYPE_H
+
+void chrpc_serial_type_tests(void);
+
+#endif
